signal-engine/requirements.txt
# Signal Engine Dependencies
# Core ML
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0

# API & Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# Drift Protocol
driftpy>=0.7.0

# Data
pyarrow>=15.0.0
python-dotenv>=1.0.0

# Technical Analysis
ta>=0.11.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0
